Version 0.0.1 Server beacons Project
Added two PWM signals that are emitted in the same time. Timer Systick
(periodic) is used for the OS. Every 250 ms the function from the
service is called. Code organization : drivers, services, applications.
The project compiles on every machine that has Keil ARM and CodeSourcery
installed with the correct path.