Version 0.0.2 Remote Control (USART) STM32-server
The state machine can change state when new UART character is received.
There are 4 characters predefined for each state : S-start, I-on, X-off,
P-stop. They need to be respected in the PC version as well in order to
respect the protocol. The UART interruption (received characters) can be
tested only in real time not in the Simulator (Keil debug).