Filter service tables and variables
All filter tables are initialized and can be automatically completed.
All the values are initialized in default as the service begin with
work. Need to update this code in order to add robustness and control
the execution. The TOA algorithm is not added because the block of 64
values doesn't introduce Enough precision. This code is not tested in
real case scenario.