Last test before the end
Test Reception and Process of the signal. These version sends debug data
to the drone in order to test the reception of the signal and the time
of process. The project is not finished and in order to validate the
communication UART and ULTRASOUND signal we need to establish a test
without the drone... TO BE CONTINUED ...